requirements.txt

python-telegram-bot==13.15
yt-dlp
openai
